requirements.txt
amqp==5.3.1
anyio==4.10.0
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==4.2.3
django-cors-headers==4.7.0
djangorestframework==3.14.0
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
kombu==5.5.4
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-telegram-bot==20.7
pytz==2025.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
